+++ spec/acceptance/classes/default_queue_spec.rb
@@ -5,6 +5,8 @@ describe 'Including class "cups"' do
     ensure_cups_is_running
   end
 
+  name = 'RSpec&Test_Printer'
+
   context 'and specifying a `default_queue`' do
     context 'when the catalog does NOT contain a `cups_queue` resource with the same name' do
       before(:all) do
@@ -13,7 +15,7 @@ describe 'Including class "cups"' do
 
       manifest = <<-EOM
         class { '::cups':
-          default_queue => 'Office'
+          default_queue => '#{name}'
         }
       EOM
 
@@ -24,16 +26,16 @@ describe 'Including class "cups"' do
 
     context 'when the catalog contains a `cups_queue` resource with the same name' do
       before(:all) do
-        add_printers(%w(BackOffice))
+        add_printers('BackOffice')
         shell('lpadmin -d BackOffice')
       end
 
       manifest = <<-EOM
         class { '::cups':
-          default_queue => 'Office'
+          default_queue => '#{name}'
         }
 
-        cups_queue { 'Office':
+        cups_queue { '#{name}':
           ensure => 'printer',
           model  => 'drv:///sample.drv/generic.ppd',
           uri    => 'lpd://192.168.2.105/binary_p1'
@@ -45,7 +47,7 @@ describe 'Including class "cups"' do
       end
 
       it 'sets the correct value' do
-        expect(shell('lpstat -d').stdout.split(/\s/)).to include('Office')
+        expect(shell('lpstat -d').stdout.split(/\s/)).to include(name)
       end
 
       it 'is idempotent' do