+++ spec/spec_helper_acceptance.rb
@@ -1,3 +1,4 @@
+require 'shellwords'
 require 'beaker-rspec'
 
 # RSpec configuration
@@ -27,40 +28,40 @@ def ensure_cups_is_running
   apply_manifest('class { "cups": }', catch_failures: true)
 end
 
-def add_printers(names)
+def add_printers(*names)
   names.each do |name|
-    shell("lpadmin -E -p #{name} -m drv:///sample.drv/generic.ppd -o printer-is-shared=false")
+    shell("lpadmin -E -p #{Shellwords.escape(name)} -m drv:///sample.drv/generic.ppd -o printer-is-shared=false")
   end
 end
 
 def add_printers_to_classes(classmembers)
-  add_printers(%w(Dummy))
+  add_printers('Dummy')
   classmembers.keys.each do |classname|
     members = classmembers[classname]
     members = %w(Dummy) if members.empty?
     members.each do |printername|
-      shell("lpadmin -E -p #{printername} -c #{classname}")
+      shell("lpadmin -E -p #{Shellwords.escape(printername)} -c #{Shellwords.escape(classname)}")
     end
-    shell("lpadmin -E -p #{classname} -o printer-is-shared=false")
+    shell("lpadmin -E -p #{Shellwords.escape(classname)} -o printer-is-shared=false")
   end
-  remove_queues(%w(Dummy))
+  remove_queues('Dummy')
 end
 
-def remove_queues(names)
-  names.each do |name|
-    shell("lpadmin -E -x #{name}", acceptable_exit_codes: [0, 1])
+def remove_queues(*names)
+  names.flatten.each do |name|
+    shell("lpadmin -E -x #{Shellwords.escape(name)}", acceptable_exit_codes: [0, 1])
   end
 end
 
 def purge_all_queues
   request = '{
-          OPERATION CUPS-Get-Printers
-          GROUP operation
-          ATTR charset attributes-charset utf-8
-          ATTR language attributes-natural-language en
-          STATUS successful-ok
-          DISPLAY printer-name
-        }'
-  result = shell("echo '#{request}' | ipptool -c ipp://localhost/ /dev/stdin", acceptable_exit_codes: [0, 1])
+    OPERATION CUPS-Get-Printers
+    GROUP operation
+    ATTR charset attributes-charset utf-8
+    ATTR language attributes-natural-language en
+    STATUS successful-ok
+    DISPLAY printer-name
+  }'
+  result = shell('ipptool -c ipp://localhost/ /dev/stdin', stdin: request, acceptable_exit_codes: [0, 1])
   remove_queues(result.stdout.split("\n")[1..-1]) if result.exit_code.zero?
 end